moves Login logic out of server file and into LoginController
