+++ server.js
@@ -3,6 +3,7 @@ const BodyareasController = require('./controllers/bodyareas-controller')
 const BodyareaLiftsController = require('./controllers/bodyarea-lifts-controller')
 const WorkoutsController = require('./controllers/workouts-controller')
 const UsersController = require('./controllers/users-controller')
+const LoginController = require('./controllers/login-controller')
 
 const express = require('express')
 const app = express()
@@ -102,28 +103,7 @@ app.get('/api/v1/users/:id', (req, res) => {
 })
 
 app.post('/login', (req, res) => {
-  const email = req.body.email
-  const password = req.body.password
-  database.raw(`SELECT id, password FROM users WHERE users.email = ?`, [email])
-  .then( data => {
-    if (data.rows.length < 1) {
-      res.sendStatus(404)
-    } else {
-      const hash = data.rows[0].password
-      const id = data.rows[0].id
-      bcrypt.compare(password, hash, (err, response) => {
-        const token = randtoken.generate(64)
-        if (response === true) {
-          database.raw(`UPDATE users SET token = ? WHERE id = ? RETURNING id, name, email, token`, [token, id])
-          .then( userInfo => {
-            res.status(200).json(userInfo.rows[0])
-          })
-        } else {
-          res.sendStatus(404)
-        }
-      })
-    }
-  })
+  LoginController.userLogin(req, res)
 })
 
 if (!module.parent) {
